sharpen:  improve clipping of sharpened line plots
The "sharpen" option can cause the y value of individual points
to reach extremely large positive or negative values.
If the resulting value exceeded INT_MAX the point was treated as
UNDEFINED rather than OUTRANGE, which would cause a gap in the plot
rather than a spike.

This commit introduces a new function map_ysharp which catches this
overflow case, and revises clip_line() to handle the full range
of possible results for (y1 - y0).

+++ src/gadgets.c
@@ -444,8 +444,8 @@ clip_line(int *x1, int *y1, int *x2, int *y2)
      * when the line passes directly through at least one corner).
      */
     count = 0;
-    dx = *x2 - *x1;
-    dy = *y2 - *y1;
+    dx = (double)(*x2) - (double)(*x1);
+    dy = (double)(*y2) - (double)(*y1);
     /* Find intersections with the x parallel bbox lines: */
     if (dy != 0) {
 	x = (clip_area->ybot - *y2) * dx / dy + *x2;	/* Test for clip_area->ybot boundary. */
